refactor word_seg model & its test
